(refactor): move `Legend` into separate file
